refactor: replace runExternalHook calls with composable
